pyproject.toml
[tool.poetry]
name = "nuanyue-fastapi"
version = "0.2.0"
description = "暖月 · 基于 FastAPI 改写的 Lin-CMS 风格后端框架"
authors = ["Pedro <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"

# --- Core Framework ---
fastapi = "^0.115.0"
starlette = "^0.48.0"
uvicorn = { extras = ["standard"], version = "^0.38.0" }

# --- Database / ORM ---
SQLAlchemy = "^2.0.44"
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"

# --- Cache / Message Queue ---
aioredis = "^2.0.1"
aio-pika = "^9.5.7"
redis = "^7.0.0"

# --- Auth / Security ---
Authlib = "^1.3.0"
passlib = "^1.7.4"
PyJWT = "^2.10.1"
bcrypt = "^5.0.0"
cryptography = "^46.0.3"

# --- Config / Settings ---
python-dotenv = "^1.1.1"
pydantic = "^2.12.3"
pydantic-settings = "^2.11.0"
PyYAML = "^6.0.3"

# --- HTTP / Async / IO ---
httpx = "^0.28.1"
aiofiles = "^24.1.0"
anyio = "^4.11.0"

# --- Utils / Logging / Tools ---
loguru = "^0.7.3"
typing-extensions = "^4.15.0"
charset-normalizer = "^3.4.4"
idna = "^3.11"

# --- Cloud / i18n ---
tencentcloud-sdk-python = "^3.0.1479"

# --- Optional ORM Plugins ---
greenlet = "^3.0.0"

# --- Misc ---
requests = "^2.32.3"

# ==============================
# Development / Docs / Test
# ==============================

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.0"
flake8 = "^7.1.1"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.5"
mypy = "^1.11.2"
watchdog = "^6.0.0"
rich = "^13.8.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
